Added cautious file checks
